Refactor font size
