+++ src/components/App/App.css
@@ -5,10 +5,11 @@
   justify-content: center;
   align-items: center;
   text-align: center;
-  background-color: aliceblue;
+  background-color:rgb(219, 219, 219)
 }
 
 .title {
+   padding-top: 1em;
   padding-bottom: 1em;
 }
