+++ src/components/ForecastCard/ForecastCard.css
@@ -2,6 +2,8 @@
   width: 33.33%;
   box-sizing: border-box;
   padding: 10px;
+  border-left: 1px solid white;
+  border-bottom: 1px solid white;
 }
 
 @media (max-width: 424px) {
