+++ src/components/ForecastContainer/ForecastContainer.css
@@ -3,4 +3,5 @@
   flex-wrap: wrap;
   align-items: center;
   justify-content: center;
+
 }
