+++ src/utils/apiCalls.js
@@ -1,4 +1,4 @@
-const getCurrentWeather = async (location) => {
+const getWeather = async (location) => {
   try {
     const response = await fetch(
       `http://localhost:3001/api/weather/${location}`
@@ -11,15 +11,5 @@ const getCurrentWeather = async (location) => {
   }
 };
 
-const getForecast = async (location) => {
-  try {
-    const response = await fetch(
-      `http://localhost:3001/api/weather/${location}/forecast`
-    );
-    const json = await response.json();
-    return json;
-  } catch (error) {
-    console.log(error);
-  }
-};
-export { getCurrentWeather, getForecast };
+
+export { getWeather };
